refactor: widgetsのui/index.tsファイルも削除
- src/widgets/posts-list/ui/index.tsを削除
- src/widgets/user-list/ui/index.tsを削除
- 上位のindex.tsファイルで直接コンポーネントをエクスポート
- 全レイヤーでFSDアーキテクチャの推奨構造に準拠
