+++ src/widgets/posts-list/index.ts
@@ -1,1 +1,1 @@
-export { PostsList } from "./ui";
+export { PostsList } from "./ui/PostsList";
