--- src/widgets/user-list/ui/index.ts
@@ -1,2 +1,0 @@
-export { UsersList } from "./UsersList";
-export { UserItem } from "./UserItem";
